refactor: rename existing `enrollment` module to `littlepay`
